[FEATURE] fixes after code review
