+++ src/Sylius/Bundle/AdminBundle/Resources/config/app/twig_hooks/shipping_method/create.yaml
@@ -46,16 +46,16 @@ sylius_twig_hooks:
                 priority: 200
             min_delivery_time_days:
                 template: '@SyliusAdmin/shipping_method/form/general/min_delivery_time_days.html.twig'
-                priority: 100
+                priority: 150
             max_delivery_time_days:
                 template: '@SyliusAdmin/shipping_method/form/general/max_delivery_time_days.html.twig'
-                priority: 0
+                priority: 125
             zone:
                 template: '@SyliusAdmin/shipping_method/form/general/zone.html.twig'
-                priority: -100
+                priority: 100
             position:
                 template: '@SyliusAdmin/shipping_method/form/general/position.html.twig'
-                priority: -200
+                priority: 0
 
         'sylius_admin.shipping_method.create.content.form.configuration':
             category:
