+++ src/Sylius/Bundle/ShippingBundle/Resources/translations/messages.en.yml
@@ -29,8 +29,6 @@ sylius:
             description: Description
             name: Name
         shipping_method:
-            min_delivery_time_days: Minimum delivery time (days)
-            max_delivery_time_days: Maximum delivery time (days)
             add_rule: Add rule
             calculator: Calculator
             category: Category
@@ -39,9 +37,11 @@ sylius:
             configuration: Configuration
             description: Description
             enabled: Enabled
+            max_delivery_time_days: Maximum delivery time (days)
             match_all_category_requirement: All units has to match the method category
             match_any_category_requirement: At least 1 unit has to match the method category
             match_none_category_requirement: None of the units have to match the method category
+            min_delivery_time_days: Minimum delivery time (days)
             name: Name
             position: Position
             rules: Rules
