+++ src/Sylius/Bundle/ShopBundle/Resources/translations/messages.pl.yml
@@ -34,11 +34,4 @@ sylius:
                 header: 'Twoje Konto'
                 order_history: 'Historia zamówień'
                 personal_information: 'Dane osobowe'
-    ui:
-        estimated_delivery_time: Dostawa
-    shop:
-        delivery_time:
-            range: '%min%-%max% dni'
-            exact: '%count% dni'
-            min_plus: 'od %min% dni'
-            max_until: 'do %max% dni'
+
