+++ src/Sylius/Bundle/ShopBundle/templates/shared/components/common/fastest_delivery_time.html.twig
@@ -1,5 +1,5 @@
-{% if text is not null %}
+{% if delivery_time is not null %}
     <div class="mt-2">
-        <small class="text-body-tertiary">{{ 'sylius.ui.estimated_delivery_time'|trans }}: {{ text }}</small>
+        <small class="text-body-tertiary">{{ 'sylius.ui.estimated_delivery_time'|trans }}: {{ delivery_time }}</small>
     </div>
 {% endif %}
