+++ src/Sylius/Bundle/ShopBundle/templates/shared/order/show/summary/table_summary/shipping/label.html.twig
@@ -7,9 +7,9 @@
     {% if order is defined and order.shipments is not empty %}
         {% set shipment = order.shipments|first %}
         {% if shipment is not null and shipment.method is not null %}
-            {% set min = attribute(shipment.method, 'minDeliveryTimeDays') %}
-            {% set max = attribute(shipment.method, 'maxDeliveryTimeDays') %}
-            {% set text %}{{ delivery.delivery_time_text(min, max) }}{% endset %}
+            {% set min = shipment.method.minDeliveryTimeDays %}
+            {% set max = shipment.method.maxDeliveryTimeDays %}
+            {% set text = delivery.delivery_time_text(min, max) %}
             {% if text is not empty %}
                 <div><small class="text-body-tertiary">{{ 'sylius.ui.estimated_delivery_time'|trans }}: {{ text }}</small></div>
             {% endif %}
